requirements.txt
pydantic>=2.0
gradio>=4.0
huggingface_hub
python-dotenv
pyyaml
